Rename file `WebsocketCloseCodes.ts` -> `IWebSocket.ts`
