+++ packages/liveblocks-core/src/types/IWebSocket.ts
